小范围优化细节

+++ app/src/main/java/com/ljyh/mei/data/model/weapi/EveryDaySongs.kt
@@ -82,7 +82,7 @@ data class EveryDaySongs(
             @SerializedName("copyright")
             val copyright: Int,
             @SerializedName("s_id")
-            val sId: Int,
+            val sId: Long,
             @SerializedName("mark")
             val mark: Long,
             @SerializedName("originCoverType")
